Fixed the errors in which synchronous methods were calling asynchronous methods.
Added ModelBinders for the nested property binding for EmailConfirmationViewModel.

Added Integration testing of two endpoints

Added paging to Filtering.

in AccountService:
fixed a bug which would allow users to sign in with incorrect passwords

in CartService:
Retrieving and Saving the cart in session would throw cycle reference errors, fixed it using JsonSerializerOptions()

in ImageService:

UpdateProductImages() method now checks the value of NewImages and RemovedImageIds lists.

in ProductService:

AddProduct wouldnt add the QuantityInStock value of the product to the base.

Implemented Skip and Take for paging in Filter method

in PriceFilter method checking LowerPriceBound on default value is removed

in SearchStringFilter removed checking the Category property containing the string.

Added GetTotalPages method for paging and SortBy method for sorting

Added unit testing for the services via mocking.

in AddProductViewModel changed Category property type to Category enum

in FilterViewModel added properties for sorting

in RegisterViewModel added property for confirming password

in ShoppingApp:
Added Views for actions and fixed error handling bugs in the Error handling middleware

Added modelbinderproviders to services in Program.cs

Removed a few validations from password requirements and refactored adding the services in proper order